bot/requirements.txt
# Telegram Bot Framework
aiogram==3.23.0
aiohttp==3.13.2

# Google Gemini AI
google-genai==1.54.0

# PostgreSQL async driver
asyncpg==0.30.0

# HTTP client for Sui RPC
httpx==0.28.1

# Configuration
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# Audio processing (for voice messages)
aiofiles==25.1.0
ffmpeg-python==0.2.0

# Utilities
tenacity==9.1.2
